+++ rsky-pds/src/sequencer/db.rs
@@ -1,11 +1,17 @@
 // based on https://github.com/bluesky-social/atproto/blob/main/packages/pds/src/sequencer/db
 
-use crate::db::migrator::{migrate_to_latest, Migration};
+use crate::db::migrator::{migrate_to_latest, Migration, MigrationSet};
 use crate::db::sqlite::Db;
 use anyhow::Result;
 use std::path::Path;
 
 pub type SequencerDb = Db;
+
+pub const SEQUENCER_DB_MIGRATIONS_SET: MigrationSet = MigrationSet {
+    shared: SEQUENCER_DB_MIGRATIONS,
+    local: &[],
+    legacy: None,
+};
 
 pub const SEQUENCER_DB_MIGRATIONS: &[Migration] = &[Migration {
     name: "001",
@@ -29,7 +35,7 @@ pub fn get_db(location: impl AsRef<Path>) -> Result<SequencerDb> {
 
 pub async fn get_migrated_db(location: impl AsRef<Path>) -> Result<SequencerDb> {
     let db = get_db(location)?;
-    migrate_to_latest(&db, SEQUENCER_DB_MIGRATIONS).await?;
+    migrate_to_latest(&db, SEQUENCER_DB_MIGRATIONS_SET).await?;
     Ok(db)
 }
 
@@ -44,7 +50,7 @@ mod tests {
             .await
             .unwrap();
         // migrating again is a no-op
-        migrate_to_latest(&db, SEQUENCER_DB_MIGRATIONS)
+        migrate_to_latest(&db, SEQUENCER_DB_MIGRATIONS_SET)
             .await
             .unwrap();
         let tables: Vec<String> = db
@@ -60,6 +66,9 @@ mod tests {
             })
             .await
             .unwrap();
-        assert_eq!(tables, ["migrations", "repo_seq"]);
+        assert_eq!(
+            tables,
+            ["kysely_migration", "kysely_migration_lock", "repo_seq"]
+        );
     }
 }